iron-fix-entrega-2

+++ iron-fix-app/src/app/pages/equipment/equipment.component.ts
@@ -4,6 +4,7 @@ import { FormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { EquipmentService } from '../../shared/services/equipment.service'
 import { AuthService } from '../../shared/services/auth.service';
 import { RouterLink } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-equipment',
@@ -63,19 +64,57 @@ export class EquipmentComponent implements OnInit {
 
   onSaveEquipment() {
     if (this.equipmentForm.invalid) return;
+
     const equipmentData = this.equipmentForm.getRawValue() as Equipment;
+    let saved: boolean;
+
     if (this.isEditing) {
-      this.equipmentService.saveEquipment(equipmentData, this.originalId);
+      saved = this.equipmentService.saveEquipment(equipmentData, this.originalId);
     } else {
-      this.equipmentService.saveEquipment(equipmentData);
+      saved = this.equipmentService.saveEquipment(equipmentData);
     }
+    if(!saved){
+      Swal.fire({
+        icon: 'error',
+        title: '¡Error!',
+        text: 'El ID ingresado ya existe',
+      });
+      return;
+    }
+
+    Swal.fire({
+      icon: 'success',
+      title: this.isEditing ? '¡Equipo actualizado!' : '¡Equipo registrado!',
+      text: this.isEditing
+        ? 'Los cambios en el equipo fueron guardados exitosamente.'
+        : 'Equipo registrado exitosamente.'
+    });
+
     this.loadEquipment();
     this.onCloseModal();
   }
 
   onDeleteEquipment(id: string) {
-    this.equipmentService.deleteEquipment(id);
-    this.loadEquipment();
+    Swal.fire({
+      title: '¿Eliminar equipo?',
+      text: `¿Estás seguro de que deseas eliminar el equipo con ID "${id}"?`,
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+      confirmButtonColor: '#ff4d4f',
+      cancelButtonColor: '#0052e0',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.equipmentService.deleteEquipment(id);
+        this.loadEquipment();
+        Swal.fire({
+          icon: 'success',
+          title: '¡Eliminado!',
+          text: `El equipo con ID "${id}" fue eliminado correctamente.`
+        });
+      }
+    });
   }
 
   onCloseModal() {
